+++ Survey2/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - In pre-process: added fail-safe if no JPG images are provided for RAW processing
 - In pre-process: added progress notification of processing
 - In calibration: calibration target image now closing after calibration process completes
+- In calibration: base tif calibration coefficients for Phantom3/Phantom4
 
 ### Changed
 - In pre-process: cleaned up output log so that extraneous information is not shown anymore